feat(init-command): refact the init command file (#77)
