+++ test/commands/init.test.js
@@ -24,6 +24,7 @@ describe('init command', () => {
         std: [
           ...loginPasswordDialog,
           // NOTICE: spinnies outputs to std.err
+          { err: 'Selecting your project' },
           { err: 'Analyzing your setup' },
           { err: 'Checking your database setup' },
           { out: 'Here are the environment variables you need to copy in your configuration file' },
@@ -43,6 +44,7 @@ describe('init command', () => {
         ],
         std: [
           // NOTICE: spinnies outputs to std.err
+          { err: 'Selecting your project' },
           { err: 'Analyzing your setup' },
           { err: 'Checking your database setup' },
           { out: 'Here are the environment variables you need to copy in your configuration file' },
